requirements.txt
python                         3.14.0
robotframework                 7.4.1
robotframework-pythonlibcore   4.4.1
robotframework-seleniumlibrary 6.8.0
selenium                       4.38.0